requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.14
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
PyJWT==2.8.0
sqlparse==0.5.1
typing-extensions==4.12.2
whitenoise==6.7.0
